requirements.txt
biopython==1.81
exceptiongroup==1.1.2
iniconfig==2.0.0
mock==5.1.0
numpy==1.25.1
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
tomli==2.0.1
